requirements.txt
torchvision
fastapi
uvicorn==0.27.1
torch
git+https://github.com/huggingface/transformers
accelerate>=1.3.0
pydantic==2.6.1
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
qwen-vl-utils[decord]==0.0.8
flash-attn
packaging
gputil
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
requests==2.31.0